move layout render into tc-ui
