+++ demo/tc-ui/edit.js
@@ -1,7 +1,12 @@
 const httpError = require('http-errors');
 const { getApiClient, getSchemaSubset } = require('./lib/tc-ui-bridge');
 const { formDataToRest, formDataToGraphQL } = require('./lib/data-conversion');
-const { handleError, renderPage } = require('./lib/renderer');
+const { Header } = require('./templates/components/header');
+const { Footer } = require('./templates/components/footer');
+const { handleError, renderPage } = require('@financial-times/tc-ui').getPageRenderer({
+	Header,
+	Footer,
+});
 const template = require('./templates/edit-page');
 
 const displayForm = async (event, apiError) => {
