+++ demo/tc-ui/templates/components/footer.jsx
