+++ packages/tc-ui/package.json
@@ -10,6 +10,7 @@
 		"react-highlight-words": "^0.16.0",
 		"showdown": "^1.9.1"
 	},
+	"main": "server.js",
 	"devDependencies": {
 		"fetch-mock": "^8.0.0",
 		"isomorphic-fetch": "^2.2.1"
